3.0.0 | Updated source location.Created hooks and axios file config in project
