+++ adm/package.json
@@ -14,6 +14,7 @@
     "@chakra-ui/react": "^2.8.1",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
+    "@tanstack/react-query": "^5.0.5",
     "axios": "^1.6.0",
     "framer-motion": "^10.16.4",
     "react": "^18.2.0",
