+++ web/src/helpers/axios.ts
@@ -1,0 +1,6 @@
+import axios from "axios";
+
+export const api = axios.create({
+    baseURL:'https://api-catalogo-pi.onrender.com'
+})
+
